requirements.txt
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
websockets==12.0
python-telegram-bot==20.7
aiogram==3.4.1
requests==2.31.0
minio==7.2.0
python-multipart==0.0.6
vk_api==11.9.9